Исправление типа данных в методе findAllFilmByDirector на Integer

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -54,7 +54,7 @@ public class FilmController {
     }
 
     @GetMapping("/director/{directorId}")
-    public List<Film> findAllFilmByDirector(@PathVariable Long directorId,
+    public List<Film> findAllFilmByDirector(@PathVariable Integer directorId,
                                             @RequestParam("sortBy") String sortBy) {
         return filmService.findAllFilmByDirector(directorId, sortBy);
     }
